API uses OrmLite

+++ IssueTracker.Data/Person.cs
@@ -1,14 +1,16 @@
 ﻿namespace IssueTracker.Data
 {
+    /// <summary>
+    /// This is a purely virtual entity
+    /// The actual data is stored as a UserAuth (provided by ServiceStack)
+    /// </summary>
     public class Person
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public int? CustomerId { get; set; }
 
         public string Name { get; set; }
-        // This isn't a public field
-        // public string Password { get; set; }
 
         public bool IsEmployee { get; set; }
     }
